requirements.txt
streamlit
pandas
numpy
scikit-learn
plotly
matplotlib
seaborn
statsmodels
prophet
xgboost
lightgbm
joblib
python-dotenv
pytest
black
flake8
pre-commit
click
